+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [**M019**] **GIS**: add new module greige inspection system.
 - [**T001**] **SPT**: enable user GAP-Transportation to edit calculated requisitions.
+- **ifs-shop-ord-vw**: add job order column.
 
 <a name="2.27.0"></a>